ASoC: Intel: ssp-common: relocate source file
Moving ssp-common source and header files from sound/soc/intel/boards
directory to sound/soc/intel/common directory and rename the source
file. Keep file content unchanged for tracking purpose.

Link: https://msgid.link/r/20240327162408.63953-2-pierre-louis.bossart@linux.intel.com
