+++ sound/soc/intel/common/sof_ssp_common.h
